Workaround Hadoop lib breakage under Jetbrains Toolbox
If Jetbrains Toolbox is started as a login item, then every time it
starts IntelliJ is firsts runs a login shell and captures current PATH
so that IntelliJ has it in its env.  (This does not occur when launching
IntelliJ as an app directly, or apparently when Toolbox is quit &
relaunched).

Hadoop libraries, when interacting with local filesystem, depend on stat
binary. They use prop stat arguments based on the OS inspection, so GNU
stat on the PATH confuses Hadoop. Perhaps this is why Brew discourages
putting GNU utils on the PATH.

This commit avoids adding GNU utils on the PATH for Toolbox.