Add open-editor
Add a script to integrate iTerm2 Semantic History with JetBrains
apps (IntelliJ IDEA, PyCharm, RustRover).

In iTerm this can be configured in Profiles » selected profile »
Advanced » Semantic History » Run command... with

    /path/to/open-editor pwd=\5 file=\1 line=\2 column=\(semanticHistory.columnNumber)

The script works by identifying whether the file to be opened is known
to some JetBrains app and which one and then asks Toolbox to open
the file in that app at requested line and position.