git cherry-pr: take base commit from upstream
Typical flow is to `git cherry-pr` onto main/master branch from the
`upstream` repo, staging the changes into own fork referenced as
`origin`. Before the changes, `origin/main` was taken as the base commit
for cherry-pick, so the `origin/main` had to be somewhat in sync with
the upstream. After this change, `main@{u}` is the base commit. Thus, as
long as the main branch's remote is `upstream`, there is no need to sync
the main branch into own fork.